BAEL-2435 Change command message package
Change the package of the command messages to coreapi.commands to signal
 that the commands are part of the API you'd share with the outside
