+++ axon/src/main/java/com/baeldung/axon/commandmodel/MessagesAggregate.java
@@ -7,8 +7,8 @@ import org.axonframework.eventsourcing.EventSourcingHandler;
 import org.axonframework.modelling.command.AggregateIdentifier;
 import org.axonframework.spring.stereotype.Aggregate;
 
-import com.baeldung.axon.commands.CreateMessageCommand;
-import com.baeldung.axon.commands.MarkReadMessageCommand;
+import com.baeldung.axon.coreapi.commands.CreateMessageCommand;
+import com.baeldung.axon.coreapi.commands.MarkReadMessageCommand;
 import com.baeldung.axon.events.MessageCreatedEvent;
 import com.baeldung.axon.events.MessageReadEvent;
